requirements.txt
streamlit>=1.32.0
plotly>=5.19.0
pandas>=2.2.0
fpdf2>=2.7.9
google-api-python-client>=2.118.0
google-auth>=2.28.1
google-auth-httplib2>=0.2.0
